# specs/001-code-learning-platform/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: AI Code Learning Platform API
  description: |
    REST API for the AI Code Learning Platform - Phase 1 MVP.

    This platform transforms AI-generated code into beginner-friendly learning materials
    through task-based code uploads, automatic analysis, document generation, and Q&A.

    Powered by Google Gemini 2.5 Flash for fast, cost-effective AI content generation.

    **Authentication**: JWT-based with HTTPOnly cookies
    **Base URL**: `/api/v1`
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://api.codelearning.com/api/v1
    description: Production server
...
paths:
...
  /tasks/{task_id}/document:
    get:
      tags: [Documents]
      summary: Get learning document
      description: Get 7-chapter learning document for task
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Document retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningDocument'
        '404':
          description: Document not found (may still be generating)
  /tasks/{task_id}/document/status:
    get:
      tags: [Documents]
      summary: Get document generation status
      description: Check status of async document generation
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Status retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [pending, in_progress, completed, failed]
                    example: in_progress
                  estimated_time_remaining:
                    type: integer
                    description: Seconds remaining (estimate)
                    example: 120
                  error:
                    type: string
                    nullable: true

  # ==================== Practice ====================
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: JWT access token in HTTPOnly cookie
  parameters:
...
    TaskId:
      name: task_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    LearningDocument:
      type: object
      properties:
        id:
          type: string
          format: uuid
        generation_status:
          type: string
          enum: [pending, in_progress, completed, failed]
        content:
          type: object
          description: 7-chapter structured content
          properties:
            chapter1:
              type: object
              properties:
                title:
                  type: string
                summary:
                  type: string
            chapter2:
              type: object
              properties:
                title:
                  type: string
                concepts:
                  type: array
                  items:
                    type: object
            # ... other chapters
        generated_at:
          type: string
          format: date-time
...